Documents/pentaMo/requirements.txt
# FastAPI Stack
fastapi>=0.111.0
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart==0.0.6
anyio>=4.8.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1

# Authentication & JWT
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
bcrypt==4.1.1

# ML & Embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
faiss-cpu==1.7.4

# Frontend & Visualization
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.3
pillow<11
packaging<24

# API & HTTP
requests==2.31.0
httpx==0.25.2
ollama==0.1.0
google-genai
huggingface-hub==0.19.4
urllib3==1.26.20

# Data Processing
numpy==1.26.4
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
unidecode==1.4.0

# Development (optional)
pytest==7.4.3
black==23.12.0
ruff==0.1.8
